{"schema":"decodex/commit/1","summary":"Split CLI control command families","authority":"XY-1115"}
