+++ apps/decodex/src/cli.rs
@@ -1,17 +1,32 @@
 mod account_commands;
+mod control_commands;
 mod docs_okf_commands;
+mod manual_commands;
 mod radar_commands;
 mod recovery_commands;
+mod research_intake_commands;
 
 use self::{
 	account_commands::AccountCommand,
+	control_commands::{
+		DiagnoseCommand, EvidenceCommand, LaneCommand, McpCommand, ProjectCommand, RunCommand,
+		ServeCommand, StatusCommand,
+	},
 	docs_okf_commands::{DocsCommand, OkfCommand},
+	manual_commands::{CommitCommand, LandCommand},
 	radar_commands::RadarCommand,
 	recovery_commands::RecoverCommand,
+	research_intake_commands::{
+		ArchiveLinearCommand, IntakeCommand, MaintenanceCommand, ResearchCommand,
+	},
 };
 #[cfg(test)]
 use self::{
 	account_commands::{AccountSubcommand, AccountUseCommand},
+	control_commands::{
+		LaneInspectCommand, LaneInterruptCommand, LaneSteerCommand, LaneSubcommand, McpSubcommand,
+		ProjectSubcommand,
+	},
 	docs_okf_commands::{
 		DocsGraphCommand, DocsSubcommand, OkfFindCommand, OkfFindFilters, OkfInitCommand,
 		OkfInitProfileArg, OkfSubcommand,
@@ -24,16 +39,19 @@ use self::{
 		StaleActiveDiagnoseCommand, StaleActiveRecoveryCommand, StaleActiveRecoverySubcommand,
 		StaleActiveReleaseCommand,
 	},
+	research_intake_commands::{
+		IntakeGoalCommand, IntakeIssuesCommand, IntakeSubcommand, ResearchCompileCommand,
+		ResearchOutcomeArg, ResearchPromoteCommand, ResearchSubcommand,
+	},
 };
 use std::{
 	fs,
 	io::{self, Read as _},
 	path::{Path, PathBuf},
-	time::Duration,
 };
 
 use clap::{
-	Args, Parser, Subcommand, ValueEnum,
+	Args, Parser, Subcommand,
 	builder::{
 		Styles,
 		styling::{AnsiColor, Effects},
@@ -43,22 +61,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
 	agent,
-	archive_hygiene::{self, ArchiveHygieneRequest},
-	maintenance::{self, MaintenanceMode, MaintenancePruneRequest, MaintenanceScope},
-	manual::{self, ManualCommitRequest, ManualLandRequest},
-	mcp::{self, McpCapabilityProfile, McpServeRequest, McpTransport},
-	orchestrator::{
-		self, DEFAULT_STEER_RESULT_WAIT_TIMEOUT, DiagnoseRequest, EvidenceRequest,
-		IssueDispatchMode, LaneInspectRequest, LaneInterruptRequest, LaneSteerReport,
-		LaneSteerRequest, RunOnceRequest, ServeRequest,
-	},
+	orchestrator::{self, IssueDispatchMode, RunOnceRequest},
 	prelude::{Result, eyre},
-	program_intake::{self, GoalIntakeCommandRequest, IssueBatchIntakeCommandRequest},
-	research_design::{
-		self, ResearchDesignCompileRequest, ResearchDesignOutcome, ResearchDesignPromoteRequest,
-		ResearchDesignRunInput,
-	},
-	runtime,
 };
 
 /// Root CLI parser for the Decodex control plane.
@@ -158,749 +162,6 @@ impl AppCommand {
 }
 
 #[derive(Debug, Args)]
-struct CommitCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Tree-change summary for the new commit message.
-	#[arg(value_name = "SUMMARY")]
-	summary: String,
-	/// Primary issue that authorizes the change. Defaults to the current issue worktree name.
-	#[arg(long, value_name = "ISSUE", conflicts_with = "manual_authority")]
-	authority: Option<String>,
-	/// Use reserved authority `manual` instead of a Linear issue.
-	#[arg(long, conflicts_with = "authority")]
-	manual_authority: bool,
-	/// Additional related issues for the commit message.
-	#[arg(long, value_name = "ISSUE")]
-	related: Vec<String>,
-	/// Mark the change as breaking.
-	#[arg(long)]
-	breaking: bool,
-}
-impl CommitCommand {
-	fn run(&self) -> Result<()> {
-		manual::run_commit(
-			self.project_config.as_path(),
-			&ManualCommitRequest {
-				summary: self.summary.clone(),
-				authority: self.authority.clone(),
-				manual_authority: self.manual_authority,
-				related: self.related.clone(),
-				breaking: self.breaking,
-			},
-		)
-	}
-}
-
-#[derive(Debug, Args)]
-struct LandCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Tree-change summary for the landed change record.
-	#[arg(value_name = "SUMMARY")]
-	summary: String,
-	/// Primary issue that authorizes the merged change. Defaults to the current issue worktree
-	/// name.
-	#[arg(long, value_name = "ISSUE", conflicts_with = "manual_authority")]
-	authority: Option<String>,
-	/// Use reserved authority `manual` instead of a Linear issue; requires `--pr`.
-	#[arg(long, conflicts_with = "authority", requires = "pr")]
-	manual_authority: bool,
-	/// Pull request URL to land. Required with `--manual-authority`; otherwise defaults to the
-	/// current review lifecycle record.
-	#[arg(long, value_name = "URL")]
-	pr: Option<String>,
-	/// Additional related issues for the landed change record.
-	#[arg(long, value_name = "ISSUE")]
-	related: Vec<String>,
-	/// Mark the landed change record as breaking.
-	#[arg(long)]
-	breaking: bool,
-}
-impl LandCommand {
-	fn run(&self) -> Result<()> {
-		manual::run_land(
-			self.project_config.as_path(),
-			&ManualLandRequest {
-				summary: self.summary.clone(),
-				authority: self.authority.clone(),
-				manual_authority: self.manual_authority,
-				pr_url: self.pr.clone(),
-				related: self.related.clone(),
-				breaking: self.breaking,
-			},
-		)
-	}
-}
-
-#[derive(Debug, Args)]
-struct RunCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Run a specific leased or queued issue by Linear identifier or tracker issue id.
-	#[arg(value_name = "ISSUE")]
-	issue: Option<String>,
-	/// Validate project loading, queue eligibility, and lane planning without tracker mutation.
-	#[arg(long)]
-	dry_run: bool,
-	/// Explain current queued candidates without preparing or dispatching a lane.
-	#[arg(long, requires = "dry_run", conflicts_with = "issue")]
-	explain: bool,
-}
-impl RunCommand {
-	fn run(&self) -> Result<()> {
-		orchestrator::run_once(RunOnceRequest {
-			config_path: self.project_config.as_path(),
-			dry_run: self.dry_run,
-			explain_queue: self.explain,
-			preferred_issue_id: self.issue.as_deref(),
-			preferred_issue_state: None,
-			preferred_initial_issue_state: None,
-			preferred_lease_acquired: false,
-			preferred_issue_claim_fd: None,
-			preferred_dispatch_slot_fd: None,
-			preferred_dispatch_slot_index: None,
-			preferred_dispatch_mode: None,
-			preferred_run_id: None,
-			preferred_attempt_number: None,
-			preferred_retry_budget_base: None,
-			preferred_workflow_snapshot: None,
-		})
-	}
-}
-
-#[derive(Debug, Args)]
-struct ServeCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Operator UI listen address.
-	#[arg(long, value_name = "ADDR", default_value_t = orchestrator::DEFAULT_OPERATOR_LISTEN_ADDRESS.to_owned())]
-	listen_address: String,
-	/// Start the local dev endpoint without polling or dispatching projects.
-	#[arg(long, hide = true)]
-	dev: bool,
-}
-impl ServeCommand {
-	fn run(&self) -> Result<()> {
-		orchestrator::run_control_plane(ServeRequest {
-			config_path: self.project_config.as_path(),
-			listen_address: &self.listen_address,
-			dev: self.dev,
-		})
-	}
-}
-
-#[derive(Debug, Args)]
-struct McpCommand {
-	#[command(subcommand)]
-	command: McpSubcommand,
-}
-impl McpCommand {
-	fn run(&self) -> Result<()> {
-		match &self.command {
-			McpSubcommand::Serve(args) => args.run(),
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct McpServeCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// MCP transport.
-	#[arg(long, value_enum, default_value_t = McpTransport::Stdio)]
-	transport: McpTransport,
-	/// Capability profile exposed by the MCP gateway. Defaults to admin for stdio and observe for
-	/// Streamable HTTP.
-	#[arg(long, value_enum)]
-	capability_profile: Option<McpCapabilityProfile>,
-	/// Streamable HTTP listen address.
-	#[arg(long, value_name = "ADDR", default_value_t = mcp::DEFAULT_MCP_HTTP_LISTEN_ADDRESS.to_owned())]
-	listen_address: String,
-	/// Trusted browser Origin for Streamable HTTP. Repeat for multiple origins.
-	#[arg(long = "allow-origin", value_name = "ORIGIN")]
-	allowed_origins: Vec<String>,
-	/// Environment variable containing the Streamable HTTP bearer token.
-	#[arg(long = "bearer-token-env", value_name = "ENV_VAR")]
-	bearer_token_env: Option<String>,
-}
-impl McpServeCommand {
-	fn run(&self) -> Result<()> {
-		mcp::serve(McpServeRequest {
-			transport: self.transport,
-			config_path: self.project_config.as_path(),
-			capability_profile: self.effective_capability_profile(),
-			listen_address: &self.listen_address,
-			allowed_origins: &self.allowed_origins,
-			bearer_token_env: self.bearer_token_env.as_deref(),
-		})
-	}
-
-	fn effective_capability_profile(&self) -> McpCapabilityProfile {
-		self.capability_profile.unwrap_or_else(|| self.transport.default_capability_profile())
-	}
-}
-
-#[derive(Debug, Args)]
-struct ProjectCommand {
-	#[command(subcommand)]
-	command: ProjectSubcommand,
-}
-impl ProjectCommand {
-	fn run(&self) -> Result<()> {
-		let state_store = runtime::open_runtime_store()?;
-
-		match &self.command {
-			ProjectSubcommand::Add(args) => {
-				let registration =
-					runtime::register_project_config(&state_store, &args.config, true)?;
-
-				if !registration.enabled() {
-					state_store.set_project_enabled(registration.service_id(), true)?;
-				}
-
-				println!(
-					"registered project {} at {}",
-					registration.service_id(),
-					registration.config_path().display()
-				);
-			},
-			ProjectSubcommand::List => {
-				let projects = state_store.list_projects()?;
-
-				if projects.is_empty() {
-					println!("No registered projects.");
-				} else {
-					for project in projects {
-						let status = if project.enabled() { "enabled" } else { "disabled" };
-
-						println!(
-							"{}\t{}\t{}",
-							project.service_id(),
-							status,
-							project.config_path().display()
-						);
-					}
-				}
-			},
-			ProjectSubcommand::Enable(args) => {
-				state_store.set_project_enabled(&args.service_id, true)?;
-
-				println!("enabled project {}", args.service_id);
-			},
-			ProjectSubcommand::Disable(args) => {
-				state_store.set_project_enabled(&args.service_id, false)?;
-
-				println!("disabled project {}", args.service_id);
-			},
-			ProjectSubcommand::Remove(args) => {
-				let removed = state_store.remove_project(&args.service_id)?;
-
-				println!(
-					"removed project {} at {}",
-					removed.service_id(),
-					removed.config_path().display()
-				);
-			},
-		}
-
-		Ok(())
-	}
-}
-
-#[derive(Debug, Args)]
-struct ProjectAddCommand {
-	/// Path to a Decodex project directory containing `project.toml` and `WORKFLOW.md`.
-	#[arg(value_name = "PROJECT_DIR")]
-	config: PathBuf,
-}
-
-#[derive(Debug, Args)]
-struct ProjectToggleCommand {
-	/// Project service id from the registered Decodex config.
-	#[arg(value_name = "SERVICE_ID")]
-	service_id: String,
-}
-
-#[derive(Debug, Args)]
-struct LaneCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	#[command(subcommand)]
-	command: LaneSubcommand,
-}
-impl LaneCommand {
-	fn run(&self) -> Result<()> {
-		match &self.command {
-			LaneSubcommand::Inspect(args) => orchestrator::print_lane_inspect(LaneInspectRequest {
-				config_path: self.project_config.as_path(),
-				issue: &args.issue,
-				run_id: args.run_id.as_deref(),
-				json: args.json,
-			}),
-			LaneSubcommand::Interrupt(args) => orchestrator::interrupt_lane(LaneInterruptRequest {
-				config_path: self.project_config.as_path(),
-				issue: &args.issue,
-				run_id: &args.run_id,
-				force: args.force,
-				reason: args.reason.as_deref(),
-				json: args.json,
-				source: "cli",
-			})
-			.map(|_report| ()),
-			LaneSubcommand::Steer(args) => args.run(self.project_config.as_path()),
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct LaneInspectCommand {
-	/// Issue identifier or local issue id to inspect.
-	#[arg(value_name = "ISSUE")]
-	issue: String,
-	/// Restrict inspection to one run id.
-	#[arg(long, value_name = "RUN_ID")]
-	run_id: Option<String>,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-}
-
-#[derive(Debug, Args)]
-struct LaneInterruptCommand {
-	/// Issue identifier or local issue id to interrupt.
-	#[arg(value_name = "ISSUE")]
-	issue: String,
-	/// Run id for the active app-server turn to interrupt.
-	#[arg(long, value_name = "RUN_ID")]
-	run_id: String,
-	/// Use hard process-kill fallback when soft interrupt is unavailable or fails.
-	#[arg(long)]
-	force: bool,
-	/// Operator-visible reason retained in local private evidence.
-	#[arg(long, value_name = "TEXT")]
-	reason: Option<String>,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-}
-
-#[derive(Debug, Args)]
-struct LaneSteerCommand {
-	/// Issue identifier or local issue id for the current lane.
-	#[arg(value_name = "ISSUE")]
-	issue: String,
-	/// Run id that must own the active turn.
-	#[arg(long, value_name = "RUN_ID")]
-	run_id: String,
-	/// Current active app-server turn id precondition.
-	#[arg(long, value_name = "TURN_ID")]
-	expected_turn_id: String,
-	/// Operator-supplied steer text to send to the active turn.
-	#[arg(long, value_name = "TEXT")]
-	message: String,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-	/// How long to wait for the active attempt to report delivery.
-	#[arg(long, value_name = "MILLISECONDS", default_value_t = default_lane_steer_wait_timeout_ms())]
-	wait_timeout_ms: u64,
-}
-impl LaneSteerCommand {
-	fn run(&self, config_path: Option<&Path>) -> Result<()> {
-		let report = orchestrator::steer_lane(LaneSteerRequest {
-			config_path,
-			project_id: None,
-			issue: &self.issue,
-			run_id: &self.run_id,
-			expected_turn_id: &self.expected_turn_id,
-			message: &self.message,
-			source: "cli",
-			wait_timeout: Duration::from_millis(self.wait_timeout_ms),
-		})?;
-
-		if self.json {
-			println!("{}", serde_json::to_string_pretty(&report)?);
-		} else {
-			print!("{}", render_lane_steer_report(&report));
-		}
-		if lane_steer_report_is_failure(&report) {
-			eyre::bail!(
-				"lane steer {}: {}",
-				report.outcome,
-				report.failure_class.as_deref().unwrap_or(&report.reason)
-			);
-		}
-
-		Ok(())
-	}
-}
-
-#[derive(Debug, Args)]
-struct StatusCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-	/// Maximum number of recent runs to display.
-	#[arg(long, value_name = "COUNT", default_value_t = orchestrator::DEFAULT_STATUS_RUN_LIMIT)]
-	limit: usize,
-	/// Refresh live tracker and pull-request observers before printing status.
-	#[arg(long)]
-	live: bool,
-}
-impl StatusCommand {
-	fn run(&self) -> Result<()> {
-		orchestrator::print_status(self.project_config.as_path(), self.json, self.limit, self.live)
-	}
-}
-
-#[derive(Debug, Args)]
-struct DiagnoseCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Emit the agent handoff index JSON instead of a one-line path summary.
-	#[arg(long)]
-	json: bool,
-	/// Maximum number of recent runs to include while generating evidence.
-	#[arg(long, value_name = "COUNT", default_value_t = orchestrator::DEFAULT_STATUS_RUN_LIMIT)]
-	limit: usize,
-}
-impl DiagnoseCommand {
-	fn run(&self) -> Result<()> {
-		orchestrator::run_diagnose(DiagnoseRequest {
-			config_path: self.project_config.as_path(),
-			json: self.json,
-			limit: self.limit,
-		})
-	}
-}
-
-#[derive(Debug, Args)]
-struct EvidenceCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Resolve this evidence readback through a registered Decodex project id.
-	#[arg(long, value_name = "SERVICE_ID")]
-	project: Option<String>,
-	/// Issue identifier or local issue id to inspect.
-	issue: String,
-	/// Restrict readback to one run id. Defaults to the latest local run for the issue.
-	#[arg(long, value_name = "RUN_ID")]
-	run_id: Option<String>,
-	/// Restrict readback to one attempt number. Defaults to the selected run attempt.
-	#[arg(long, value_name = "NUMBER")]
-	attempt: Option<i64>,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-	/// Include full structured payload values instead of compact payload summaries only.
-	#[arg(long)]
-	include_payload: bool,
-}
-impl EvidenceCommand {
-	fn run(&self) -> Result<()> {
-		orchestrator::print_private_evidence(EvidenceRequest {
-			config_path: self.project_config.as_path(),
-			project_id: self.project.as_deref(),
-			issue: &self.issue,
-			run_id: self.run_id.as_deref(),
-			attempt_number: self.attempt,
-			json: self.json,
-			include_payload: self.include_payload,
-		})
-	}
-}
-
-#[derive(Debug, Args)]
-struct IntakeCommand {
-	#[command(subcommand)]
-	command: IntakeSubcommand,
-}
-impl IntakeCommand {
-	fn run(&self) -> Result<()> {
-		match &self.command {
-			IntakeSubcommand::Goal(args) => args.run(),
-			IntakeSubcommand::Issues(args) => args.run(),
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct IntakeGoalCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Registered Decodex service id to intake against.
-	#[arg(long, value_name = "SERVICE_ID", conflicts_with = "config")]
-	project: Option<String>,
-	/// Read tracker state and print the deterministic goal-intake report without mutation.
-	#[arg(long, conflicts_with = "apply", required_unless_present = "apply")]
-	dry_run: bool,
-	/// Create or update generated normal Linear issues and persist local Program Intake state.
-	#[arg(long, conflicts_with = "dry_run")]
-	apply: bool,
-	/// Existing Linear issue whose team and startable state should anchor generated issues.
-	#[arg(long, value_name = "ISSUE")]
-	team_issue: Option<String>,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-	/// Accepted Decision Contract identifier to materialize.
-	#[arg(value_name = "CONTRACT_ID")]
-	contract_id: String,
-}
-impl IntakeGoalCommand {
-	fn run(&self) -> Result<()> {
-		let report = program_intake::run_goal_intake_command(GoalIntakeCommandRequest {
-			config_path: self.project_config.as_path(),
-			project_id: self.project.as_deref(),
-			contract_id: &self.contract_id,
-			team_issue_identifier: self.team_issue.as_deref(),
-			dry_run: self.dry_run,
-			apply: self.apply,
-		})?;
-
-		if self.json {
-			println!("{}", serde_json::to_string_pretty(&report)?);
-		} else {
-			print!("{}", program_intake::render_goal_intake_report(&report));
-		}
-
-		Ok(())
-	}
-}
-
-#[derive(Debug, Args)]
-struct IntakeIssuesCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Registered Decodex service id to intake against.
-	#[arg(long, value_name = "SERVICE_ID", conflicts_with = "config")]
-	project: Option<String>,
-	/// Read tracker state and print the deterministic intake report without local persistence.
-	#[arg(long, conflicts_with = "apply", required_unless_present = "apply")]
-	dry_run: bool,
-	/// Persist local runtime Program Intake records for direct Program dispatch.
-	#[arg(long, conflicts_with = "dry_run")]
-	apply: bool,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-	/// Existing Linear issue identifiers to intake into an internal Execution Program.
-	#[arg(value_name = "ISSUE")]
-	#[arg(required = true)]
-	issues: Vec<String>,
-}
-impl IntakeIssuesCommand {
-	fn run(&self) -> Result<()> {
-		let report =
-			program_intake::run_issue_batch_intake_command(IssueBatchIntakeCommandRequest {
-				config_path: self.project_config.as_path(),
-				project_id: self.project.as_deref(),
-				issue_identifiers: self.issues.clone(),
-				dry_run: self.dry_run,
-				persist: self.apply,
-			})?;
-
-		if self.json {
-			println!("{}", serde_json::to_string_pretty(&report)?);
-		} else {
-			print!("{}", program_intake::render_issue_batch_intake_report(&report));
-		}
-
-		Ok(())
-	}
-}
-
-#[derive(Debug, Args)]
-struct ResearchCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	#[command(subcommand)]
-	command: ResearchSubcommand,
-}
-impl ResearchCommand {
-	fn run(&self) -> Result<()> {
-		match &self.command {
-			ResearchSubcommand::Compile(args) => args.run(self.project_config.as_path()),
-			ResearchSubcommand::Promote(args) => args.run(self.project_config.as_path()),
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct ResearchCompileCommand {
-	/// Structured research/design input JSON. Use `-` to read from stdin.
-	#[arg(long, value_name = "JSON")]
-	input: Option<PathBuf>,
-	/// Natural-language research/design intent for minimal intake.
-	#[arg(long, value_name = "TEXT", conflicts_with = "input")]
-	intent: Option<String>,
-	/// Source tracker issue identifier to link to the contract.
-	#[arg(long, value_name = "ISSUE")]
-	source_issue: Option<String>,
-	/// Outcome for minimal natural-language intake.
-	#[arg(long, value_enum, default_value = "not-decision-ready")]
-	outcome: ResearchOutcomeArg,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-}
-impl ResearchCompileCommand {
-	fn run(&self, config_path: Option<&Path>) -> Result<()> {
-		let input = self.research_input()?;
-		let report =
-			research_design::run_compile(ResearchDesignCompileRequest { config_path, input })?;
-
-		if self.json {
-			println!("{}", serde_json::to_string_pretty(&report)?);
-		} else {
-			println!(
-				"research compile {}: contract={} status={} ready_after_promotion={} authority={}",
-				research_outcome_label(report.outcome),
-				report.contract_id,
-				report.contract_status.as_str(),
-				report.issue_generation_ready_after_promotion,
-				report.execution_authority_granted
-			);
-			println!("{}", report.feedback);
-		}
-
-		Ok(())
-	}
-
-	fn research_input(&self) -> Result<ResearchDesignRunInput> {
-		match (&self.input, &self.intent) {
-			(Some(path), None) => read_research_input(path),
-			(None, Some(intent)) => Ok(ResearchDesignRunInput::from_intent(
-				intent.clone(),
-				self.source_issue.clone(),
-				self.outcome.into(),
-			)),
-			(None, None) => {
-				eyre::bail!("research compile requires either --input <JSON> or --intent <TEXT>.")
-			},
-			(Some(_), Some(_)) => {
-				eyre::bail!("research compile accepts --input or --intent, not both.")
-			},
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct ResearchPromoteCommand {
-	/// Decision Contract identifier to promote.
-	#[arg(value_name = "CONTRACT_ID")]
-	contract_id: String,
-	/// Actor accepting the contract.
-	#[arg(long, value_name = "TEXT", default_value = "operator")]
-	accepted_by: String,
-	/// Acceptance source, usually conversation or runtime policy.
-	#[arg(long, value_name = "TEXT", default_value = "conversation")]
-	acceptance_source: String,
-	/// RFC3339 acceptance timestamp. Defaults to current UTC time.
-	#[arg(long, value_name = "RFC3339")]
-	accepted_at: Option<String>,
-	/// Optional acceptance reason.
-	#[arg(long, value_name = "TEXT")]
-	reason: Option<String>,
-	/// Emit structured JSON instead of human-readable text.
-	#[arg(long)]
-	json: bool,
-}
-impl ResearchPromoteCommand {
-	fn run(&self, config_path: Option<&Path>) -> Result<()> {
-		let report = research_design::run_promote(ResearchDesignPromoteRequest {
-			config_path,
-			contract_id: &self.contract_id,
-			accepted_by: &self.accepted_by,
-			accepted_at: self.accepted_at.as_deref(),
-			acceptance_source: &self.acceptance_source,
-			promotion_reason: self.reason.clone(),
-		})?;
-
-		if self.json {
-			println!("{}", serde_json::to_string_pretty(&report)?);
-		} else {
-			println!(
-				"research promote: contract={} status={} authority={} ready={}",
-				report.contract_id,
-				report.contract_status.as_str(),
-				report.execution_authority_granted,
-				report.ready_for_issue_shaping
-			);
-		}
-
-		Ok(())
-	}
-}
-
-#[derive(Debug, Args)]
-struct ArchiveLinearCommand {
-	#[command(flatten)]
-	project_config: ProjectConfigArgs,
-	/// Repo label scope to inspect, for example `repo:decodex`.
-	#[arg(long = "repo-label", value_name = "LABEL", required = true)]
-	repo_labels: Vec<String>,
-	/// Archive only issues last updated more than this many days ago.
-	#[arg(long, value_name = "DAYS", default_value_t = 30)]
-	older_than_days: u32,
-	/// Perform the archive mutation. Omit this flag for the dry-run candidate report.
-	#[arg(long)]
-	execute: bool,
-}
-impl ArchiveLinearCommand {
-	fn run(&self) -> Result<()> {
-		archive_hygiene::run(
-			self.project_config.as_path(),
-			&ArchiveHygieneRequest {
-				repo_labels: self.repo_labels.clone(),
-				older_than_days: self.older_than_days,
-				execute: self.execute,
-			},
-		)
-	}
-}
-
-#[derive(Debug, Args)]
-struct MaintenanceCommand {
-	#[command(subcommand)]
-	command: MaintenanceSubcommand,
-}
-impl MaintenanceCommand {
-	fn run(&self) -> Result<()> {
-		match &self.command {
-			MaintenanceSubcommand::Prune(args) => args.run(),
-		}
-	}
-}
-
-#[derive(Debug, Args)]
-struct MaintenancePruneCommand {
-	/// Report candidates without applying retention changes. This is the default mode.
-	#[arg(long, conflicts_with = "apply")]
-	dry_run: bool,
-	/// Apply safe file retention, state-aware runtime compaction, and WAL checkpointing.
-	#[arg(long, conflicts_with = "dry_run")]
-	apply: bool,
-	/// Emit the maintenance report as JSON.
-	#[arg(long)]
-	json: bool,
-}
-impl MaintenancePruneCommand {
-	fn run(&self) -> Result<()> {
-		let mode = if self.apply { MaintenanceMode::Apply } else { MaintenanceMode::DryRun };
-
-		maintenance::run_prune_command(MaintenancePruneRequest {
-			mode,
-			scope: MaintenanceScope::Full,
-			json: self.json,
-		})
-	}
-}
-
-#[derive(Debug, Args)]
 struct ProbeCommand {
 	/// Override the expected app-server transport during probing.
 	#[arg(value_name = "TRANSPORT", default_value = "stdio://")]
@@ -975,17 +236,6 @@ impl From<AttemptDispatchMode> for IssueDispatchMode {
 	}
 }
 
-impl From<ResearchOutcomeArg> for ResearchDesignOutcome {
-	fn from(value: ResearchOutcomeArg) -> Self {
-		match value {
-			ResearchOutcomeArg::DecisionReady => Self::DecisionReady,
-			ResearchOutcomeArg::NotDecisionReady => Self::NotDecisionReady,
-			ResearchOutcomeArg::Blocked => Self::Blocked,
-			ResearchOutcomeArg::NeedsHumanDecision => Self::NeedsHumanDecision,
-		}
-	}
-}
-
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub(crate) enum AttemptDispatchMode {
@@ -1055,116 +305,6 @@ enum Command {
 	/// Run one daemon-planned attempt from a structured request.
 	#[command(name = "_attempt", hide = true)]
 	Attempt(AttemptCommand),
-}
-
-#[derive(Debug, Subcommand)]
-enum ProjectSubcommand {
-	/// Register or refresh one Decodex project config and enable it.
-	Add(ProjectAddCommand),
-	/// List registered local projects.
-	List,
-	/// Enable one registered project for `decodex serve`.
-	Enable(ProjectToggleCommand),
-	/// Disable one registered project for `decodex serve`.
-	Disable(ProjectToggleCommand),
-	/// Remove one registered project from the local registry.
-	Remove(ProjectToggleCommand),
-}
-
-#[derive(Debug, Subcommand)]
-enum McpSubcommand {
-	/// Serve Decodex MCP protocol primitives.
-	Serve(McpServeCommand),
-}
-
-#[derive(Debug, Subcommand)]
-enum LaneSubcommand {
-	/// Inspect one local lane by issue identifier or tracker issue id.
-	Inspect(LaneInspectCommand),
-	/// Soft-interrupt an active app-server turn, with optional hard fallback.
-	Interrupt(LaneInterruptCommand),
-	/// Send operator-supplied text to an active steerable turn.
-	Steer(LaneSteerCommand),
-}
-
-#[derive(Debug, Subcommand)]
-enum ResearchSubcommand {
-	/// Compile bounded research/design input into a latent Decision Contract.
-	Compile(ResearchCompileCommand),
-	/// Promote an accepted Decision Contract into execution authority.
-	Promote(ResearchPromoteCommand),
-}
-
-#[derive(Debug, Subcommand)]
-enum IntakeSubcommand {
-	/// Dry-run or apply a promoted Decision Contract as normal issue-backed goal intake.
-	Goal(IntakeGoalCommand),
-	/// Dry-run or persist existing Linear issues as an internal program intake batch.
-	Issues(IntakeIssuesCommand),
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, ValueEnum)]
-#[value(rename_all = "kebab-case")]
-enum ResearchOutcomeArg {
-	DecisionReady,
-	NotDecisionReady,
-	Blocked,
-	NeedsHumanDecision,
-}
-
-#[derive(Debug, Subcommand)]
-enum MaintenanceSubcommand {
-	/// Inspect or apply local Decodex storage retention.
-	Prune(MaintenancePruneCommand),
-}
-
-fn default_lane_steer_wait_timeout_ms() -> u64 {
-	u64::try_from(DEFAULT_STEER_RESULT_WAIT_TIMEOUT.as_millis()).unwrap_or(10_000)
-}
-
-fn lane_steer_report_is_failure(report: &LaneSteerReport) -> bool {
-	matches!(report.outcome.as_str(), "rejected" | "failed" | "timed_out" | "fallback")
-}
-
-fn research_outcome_label(outcome: ResearchDesignOutcome) -> &'static str {
-	match outcome {
-		ResearchDesignOutcome::DecisionReady => "decision-ready",
-		ResearchDesignOutcome::NotDecisionReady => "not-decision-ready",
-		ResearchDesignOutcome::Blocked => "blocked",
-		ResearchDesignOutcome::NeedsHumanDecision => "needs-human-decision",
-	}
-}
-
-fn render_lane_steer_report(report: &LaneSteerReport) -> String {
-	format!(
-		"lane steer {}: issue={} run_id={} attempt={} expected_turn_id={} current_turn_id={} response_turn_id={} failure_class={} audit_record_id={} delivery_status={}\n",
-		report.outcome,
-		report.issue_identifier.as_deref().unwrap_or(&report.issue_id),
-		report.run_id,
-		report.attempt_number,
-		report.expected_turn_id,
-		report.current_turn_id.as_deref().unwrap_or("none"),
-		report.response_turn_id.as_deref().unwrap_or("none"),
-		report.failure_class.as_deref().unwrap_or("none"),
-		report.audit_record_id,
-		report.delivery_status
-	)
-}
-
-fn read_research_input(path: &Path) -> Result<ResearchDesignRunInput> {
-	let raw = if path == Path::new("-") {
-		let mut raw = String::new();
-
-		io::stdin().read_to_string(&mut raw)?;
-
-		raw
-	} else {
-		fs::read_to_string(path)?
-	};
-
-	serde_json::from_str(&raw).map_err(|error| {
-		eyre::eyre!("Failed to parse research input `{}`: {error}", path.display())
-	})
 }
 
 fn read_attempt_request(request: &str) -> Result<AttemptRequest> {
